# SME-Connect/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Biashara Ndogo Association of Kenya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: county
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: registerMember
      tags: [members]
      summary: Register a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Validation error
...
  /sectors/{id}:
    get:
      operationId: getSector
      tags: [sectors]
      summary: Get sector profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sector profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
        "404":
          description: Not found
...
  /partners/applications:
    post:
      operationId: createPartnerApplication
      tags: [partners]
      summary: Submit a partnership application for BNAK review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerApplicationInput"
      responses:
        "202":
          description: Application received and pending review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerApplication"
        "400":
          description: Invalid application or anti-bot verification
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        category:
          type: string
        county:
          type: string
        constituency:
          type: string
        ward:
          type: string
        ageBracket:
          type: string
        gender:
          type: string
        nationalId:
          type: string
        estate:
          type: string
        businessLocation:
          type: string
        businessRegistrationNumber:
          type: string
        kraPin:
          type: string
        monthlyTransactionBracket:
          type: string
        mpesaNumber:
          type: string
        membershipNumber:
          type: string
        status:
          type: string
          enum: [active, pending, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - phone
        - businessName
        - businessType
        - category
        - county
        - constituency
        - ward
        - membershipNumber
        - status
        - createdAt
    MemberInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        businessName:
          type: string
          minLength: 2
        businessType:
          type: string
        category:
          type: string
        county:
          type: string
        constituency:
          type: string
        ward:
          type: string
        ageBracket:
          type: string
        gender:
          type: string
        nationalId:
          type: string
        estate:
          type: string
        businessLocation:
          type: string
        businessRegistrationNumber:
          type: string
        kraPin:
          type: string
        monthlyTransactionBracket:
          type: string
        mpesaNumber:
          type: string
        verificationToken:
          type: string
        verificationAnswer:
          type: string
        formStartedAt:
          type: integer
        website:
          type: string
      required:
        - fullName
        - email
        - phone
        - businessName
        - businessType
        - category
        - county
        - constituency
        - ward
        - verificationToken
        - verificationAnswer
        - formStartedAt
...
    PartnerApplication:
      type: object
      properties:
        id:
          type: integer
        organisationName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: ["string", "null"]
        partnershipType:
          type: string
          enum: [Strategic Partner, Programme Partner, Implementing Partner, Funding Partner, Sponsor, Technical Partner, Knowledge Partner, Market Partner]
        organisationSummary:
          type: string
        proposedContribution:
          type: string
        countyOrCoverage:
          type: string
        status:
          type: string
          enum: [pending_review, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, organisationName, contactName, email, phone, partnershipType, organisationSummary, proposedContribution, countyOrCoverage, status, createdAt]
    PartnerApplicationInput:
      type: object
      properties:
        organisationName:
          type: string
          minLength: 2
        contactName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 9
        website:
          type: string
          format: uri
        partnershipType:
          type: string
          enum: [Strategic Partner, Programme Partner, Implementing Partner, Funding Partner, Sponsor, Technical Partner, Knowledge Partner, Market Partner]
        organisationSummary:
          type: string
          minLength: 20
          maxLength: 1500
        proposedContribution:
          type: string
          minLength: 20
          maxLength: 1500
        countyOrCoverage:
          type: string
          minLength: 2
        verificationToken:
          type: string
        verificationAnswer:
          type: string
        formStartedAt:
          type: integer
        honeypot:
          type: string
      required: [organisationName, contactName, email, phone, partnershipType, organisationSummary, proposedContribution, countyOrCoverage, verificationToken, verificationAnswer, formStartedAt]
...
    Sector:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        totalSMEs:
          type: integer
        annualContribution:
          type: string
        keyProducts:
          type: array
          items:
            type: string
        challenges:
          type: array
          items:
            type: string
        opportunities:
          type: array
          items:
            type: string
        counties:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - imageUrl
        - totalSMEs
        - annualContribution
        - keyProducts
        - challenges
        - opportunities
        - counties
...