# SME-Connect/lib/api-zod/src/generated/types/member.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biashara Ndogo Association of Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberStatus } from './memberStatus';

export interface Member {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  businessName: string;
  businessType: string;
  category: string;
  county: string;
  constituency: string;
  ward: string;
  ageBracket?: string;
  gender?: string;
  nationalId?: string;
  estate?: string;
  businessLocation?: string;
  businessRegistrationNumber?: string;
  kraPin?: string;
  monthlyTransactionBracket?: string;
  mpesaNumber?: string;
  membershipNumber: string;
  status: MemberStatus;
  createdAt: Date;
}